+++ kubejs/server_scripts/Greg Recipies/Recipe Changes/create.js
@@ -257,31 +257,6 @@ ServerEvents.recipes(event => {
             'gtceu:andesite_alloy_rod'
         ]
     )
-//Create casings
-    function casingass(id, plate, log, output){
-        event.recipes.gtceu.assembler(id)
-            .itemInputs(plate, log)
-            .itemOutputs(output)
-            .duration(50)
-            .EUt(16)
-            .circuit(6);
-        }
-    function casingcraft(output, plate, log){
-        event.shaped(
-            Item.of(output, 2),
-            [
-                'ACA',
-                'ABA',
-                'ADA'
-            ],
-            {
-                A: plate,
-                B: log,
-                C: '#forge:tools/hammers',
-                D: '#forge:tools/wrenches',
-            }
-            )
-        }
 //Create pipes
     function createpipe(output, metal){
         event.shaped(
@@ -642,28 +617,6 @@ ServerEvents.recipes(event => {
     funnel('2x create:brass_funnel', 'gtceu:brass_plate', 'create:electron_tube', 'minecraft:dried_kelp')
     funnel('2x create:andesite_tunnel', 'gtceu:andesite_alloy_plate', 'gtceu:small_andesite_alloy_gear', 'create:belt_connector')
     funnel('2x create:brass_tunnel', 'gtceu:brass_plate', 'create:electron_tube', 'create:belt_connector')
-
-    casingass('andesite_casing_log', '6x gtceu:andesite_alloy_plate', '#forge:stripped_logs', '2x create:andesite_casing')
-    casingass('andesite_casing_wood', '6x gtceu:andesite_alloy_plate', '#forge:stripped_wood', '2x create:andesite_casing')
-    casingass('brass_casing_log', '6x gtceu:brass_plate', '#forge:stripped_logs', '2x create:brass_casing')
-    casingass('brass_casing_wood', '6x gtceu:brass_plate', '#forge:stripped_wood', '2x create:brass_casing')
-    casingass('copper_casing_log', '6x gtceu:copper_plate', '#forge:stripped_logs', '2x create:copper_casing')
-    casingass('copper_casing_wood', '6x gtceu:copper_plate', '#forge:stripped_wood', '2x create:copper_casing')
-    casingass('steel_casing_log', '6x gtceu:steel_plate', 'kubejs:stripped_treated_wood_log', '2x tfmg:steel_casing')
-    casingass('steel_casing_wood', '6x gtceu:steel_plate', 'kubejs:stripped_treated_wood', '2x tfmg:steel_casing')
-    //casingass('electrum_casing_log', '6x gtceu:electrum_plate', 'kubejs:stripped_treated_wood_log', '2x kubejs:electrum_casing')
-    //casingass('electrum_casing_wood', '6x gtceu:electrum_plate', 'kubejs:stripped_treated_wood', '2x kubejs:electrum_casing')
-
-    casingcraft('create:andesite_casing', 'gtceu:andesite_alloy_plate', '#forge:stripped_logs')
-    casingcraft('create:andesite_casing', 'gtceu:andesite_alloy_plate', '#forge:stripped_wood')
-    casingcraft('create:brass_casing', 'gtceu:brass_plate', '#forge:stripped_logs')
-    casingcraft('create:brass_casing', 'gtceu:brass_plate', '#forge:stripped_wood')
-    casingcraft('create:copper_casing', 'gtceu:copper_plate', '#forge:stripped_logs')
-    casingcraft('create:copper_casing', 'gtceu:copper_plate', '#forge:stripped_wood')
-    casingcraft('tfmg:steel_casing', 'gtceu:steel_plate', 'kubejs:stripped_treated_wood_log')
-    casingcraft('tfmg:steel_casing', 'gtceu:steel_plate', 'kubejs:stripped_treated_wood')
-    //casingcraft('kubejs:electrum_casing', 'gtceu:electrum_plate', 'kubejs:stripped_treated_wood_log')
-    //casingcraft('kubejs:electrum_casing', 'gtceu:electrum_plate', 'kubejs:stripped_treated_wood')
 
     mechcreate('create:encased_fan', 'gtceu:andesite_alloy_rotor', 'create:shaft')
     mechcreate('create:mechanical_press', 'gtceu:double_iron_plate', 'create:shaft')