+++ kubejs/startup_scripts/gtceu/Multiblocks/hydroponic_farm.js
@@ -12,7 +12,7 @@ GTCEuStartupEvents.registry('gtceu:machine', event => {
     event.create('hydroponic_farm', 'multiblock')
         .rotationState(RotationState.NON_Y_AXIS)
         .recipeType('hydroponic_farm')
-        .appearanceBlock(GTBlocks.CASING_STEEL_SOLID)
+        .appearanceBlock('kubejs:agri_casing')
         .pattern(definition => FactoryBlockPattern.start()
             .aisle('CCCCCCCCCCC', 'CCCCCCCCCCC', 'CGGGGGGGGCC', 'CCCCCCCCCCC')
             .aisle('EEEEEEEEEEE', 'EDDDDDDDDEE', 'G        PI', 'EGGGGGGGGEE').setRepeatable(1, 8)
@@ -22,8 +22,8 @@ GTCEuStartupEvents.registry('gtceu:machine', event => {
             .where('D', Predicates.blocks('minecraft:dirt'))
             .where('P', Predicates.blocks(GTBlocks.CASING_STEEL_PIPE.get()))
             .where('G', Predicates.blocks(GTBlocks.CASING_TEMPERED_GLASS.get()))
-            .where('E', Predicates.blocks(GTBlocks.CASING_STEEL_SOLID.get()))
-            .where('C', Predicates.blocks(GTBlocks.CASING_STEEL_SOLID.get())
+            .where('E', Predicates.blocks('kubejs:agri_casing'))
+            .where('C', Predicates.blocks('kubejs:agri_casing')
                     .or(Predicates.abilities(PartAbility.MAINTENANCE).setExactLimit(1))
                     .or(Predicates.abilities(PartAbility.IMPORT_FLUIDS).setMinGlobalLimited(1).setMaxGlobalLimited(2))
                     .or(Predicates.abilities(PartAbility.INPUT_ENERGY).setMinGlobalLimited(1).setMaxGlobalLimited(2))
@@ -32,7 +32,7 @@ GTCEuStartupEvents.registry('gtceu:machine', event => {
             .build()
         )
         .workableCasingRenderer(
-            'gtceu:block/casings/solid/machine_casing_solid_steel',
+            'gtceu:block/casings/agri_casing',
             'gtceu:block/multiblock/implosion_compressor', false
         )
 })